python/digitize_service/requirements.txt
# EKGQuest Digitization Service - Dependencies
# Install with: pip install -r requirements.txt

flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
anthropic>=0.39.0
